Maintain DB code separately from upstream
