+++ chef/cookbooks/bcpc/recipes/nova-head.rb
@@ -380,6 +380,7 @@ if extended_apis['enabled']
   %w(
     /usr/lib/python3/dist-packages/nova/api/openstack/compute/bcpc
     /usr/lib/python3/dist-packages/nova/compute/bcpc
+    /usr/lib/python3/dist-packages/nova/db/main/bcpc
   ).each do |path|
     directory path do
       action :create
@@ -396,10 +397,11 @@ if extended_apis['enabled']
     'server_properties_policy.py' => 'policies/server_properties.py',
     'bcpc_compute_init.py' => 'compute/bcpc/__init__.py',
     'bcpc_compute_api.py' => 'compute/bcpc/api.py',
+    'bcpc_db_init.py' => 'db/main/bcpc/__init__.py',
+    'bcpc_db_api.py' => 'db/main/bcpc/api.py',
     'bcpc_routes_init.py' => 'api/openstack/compute/bcpc/__init__.py',
     'bcpc_routes.py' => 'api/openstack/compute/bcpc/bcpc_routes.py',
     # Files that we are patching over
-    'db_main_api.py' => 'db/main/api.py',
     'instance.py' => 'objects/instance.py',
     'policies_init.py' => 'policies/__init__.py',
   }
